Split generator page into GeneratorComponent and ImageDescriptor components
